Merged PR 57: Adds NUnit project to assert results of NLU services
Adds NUnit project to assert results of NLU services

Adds the LanguageUnderstanding.Evaluator project, which generates a test case for each labeled utterance in the test case file and compares it to the corresponding (by array index) utterance in the test results file. The project is configured to run a hard-coded set of expected and actual utterances (which constitute the unit tests), but you can easily override this configuration by dropping in an `appsettings.local.json` file. The entire process should be easy to automate:
1. Generate test results from labeled utterances file for NLU service
2. Generate `appsettings.local.json` pointing to input and output labeled utterances files in LanguageUnderstanding.Evaluator project
3. Run `dotnet test` for the LanguageUnderstanding.Evaluator project

Fixes #161

Related work items: #161