Merged PR 100: Adjust unit tests with better descriptions
Adjust unit tests with better descriptions

The entity unit tests failures were not always descriptive for why a failure occurred. This change makes the tests more granular to specify which entity is being tested, so when the failure occurs it is easier to see why.

Also, loosens the equality comparison for the regular entity check case, so that either match text or entity value may be equal to the match text or entity value of the comparand for the test to pass. Added a strict entity value check to make sure the entity values are strictly compared for semantic equality purposes.

Fixes #201
Fixes #217

Related work items: #201, #217