Merged PR 107: Fix NLU services to allow training to existing app / bot
Fix NLU services to allow training to existing app / bot

Adds two operations to Lex, one to check if the bot already exists and another to check if the bot alias already exists. If so, the PutBot and PutBotAlias operations will be skipped, respectively.

Also, while validating task #182 for LUIS, it was easier to test if it was possible to import to a specific LUIS app if we were able to overwrite the "saved" app settings. Thus, switching the order of 'appsettings.local.json' and 'appsettings.luis.json' improved the developer experience.

Fixes #182

Related work items: #182