Merged PR 156: Remove entity types in favor of service-specific settings
Remove entity types in favor of service-specific settings

The implementations of Lex and LUIS diverged significantly after introducing the option to include templates. This just includes the template as part of the former "entityTypes" file, which is now a mapping file for LUIS and an entity types configuration for Lex.