Merged PR 164: Adds support for plugin NLU service
Adds support for plugin NLU service

This commit enables external implementations of the `INLUServiceFactory` through the System.Composition namespace. Users can supply a path to the service assembly, or the tool will search the .NET tools path where `dotnet-nlu` is installed for an assembly matching `dotnet-nlu-*`, where `*` is replaced with the service identifier supplied on the command line.