Merged PR 166: Replace `recordingID` with `speechFile` in test utterances
Replace `recordingID` with `speechFile` in test utterances

While documenting, I realized that there was no reason for the indirection incurred by the `recordingID`, when we could just specify the speech file path. This would be supportive of other NLU services, than may not require the WAV 16KHz format. Now, the `-d` option is truly optional, and is applied as a prefix to the `speechFile` property in the labeled test utterances if supplied on the command line.