Adds documentation for assigning Azure resource for LUIS
An important feature for LUIS is the ability to assign an Azure resource to the LUIS app that gets created for training. This allows us to create a fresh LUIS app that will only be used within the scope of the CI build, and will not be subject to the authoring key quota.