Adds project to support LUIS v3
In order to keep things stable in the near term, we will maintain two versions of the LUIS NLU service provider, one for LUIS v2 and the other for LUIS v3. For the most part, the LUIS v3 library is a copy paste of LUIS v2, and some minor modifications to work with the latest API changes.

The LUIS v3 library supports the new runtime API, including the ability to use "query" instead of "text" for the test utterance, and the ability to add extra JSON properties to your request, such as "externalEntities", "options" and "dynamicLists"

There is a remaining TODO to figure out what to do with entity value resolution, as the entity values in LUIS v3 follow a less consistent pattern than in v2. Follow issue #27 for progress on this issue.

Fixes #39