Removes LUIS and Lex project references in NLU.DevOps.CommandLine
Previously, the NLU.DevOps.Luis and NLU.DevOps.Lex were direct project references in NLU.DevOps.CommandLine. This caused a problem for NLU.DevOps.LuisV3, as we could not enable it to be available by default (or at all) since the LUIS SDK dependency would conflict with the version used in NLU.DevOps.Luis.

This PR fixes up the solution to copy the build output from the LUIS, Lex, and LUIS v3 projects into provider specific folders, and uses the `AppDomain.CurrentDomain.AssemblyResolve` event to pull in dependencies for a specific NLU service provider at runtime.

The side-effect of this PR is that it no longer allows you to use parallel builds in Visual Studio (or with the MSBuild command line), since the NLU.DevOps.CommandLine build would be partially evaluated before the build output for the NLU service provider projects were copied to their respective build output (bin) folders.

Fixes #606