Switches from custom mocks to Moq
Moq does a lot of the work that we were doing to capture calls to I[Lex|Luis][Train|Test]Client, and includes some additional features for further guarantees that expected arguments are being used. This changeset swaps out the custom mocks we built for Moq, resulting in over 500 lines of code being removed.