Adds `EntityResolution` value and removes `INLUQuery`
Adds `Entity.EntityResolution` so that NLU.DevOps users can assert the correct resolution value for LUIS entities. In a future PR, we may add support to allow users to specify an arbitrary JSON assert function.

Changes `Entity.EntityValue` to `JToken` so that we can have arbitrary entity values, e.g., for LUIS composite entities.

The main reason for introducing `INLUQuery` was to avoid taking a dependency on `Newtonsoft.Json` in the `NLU.DevOps.Models` project. However, since we're taking the dependency for `Entity.EntityResolution`, it no longer makes sense to include this abstraction layer.

Fixes #27