Simplify entity model by removing resolution
Removes entity resolution from Entity model and all associated usages. The `entityResolution` option for comparison was overly specific to LUIS, and a similar effect can be achieved by putting the putting the `resolution` values as is into the `entityValue` field, rather than trying to extract a singular value.

Note - this is a breaking change for LUIS v2, as we no longer lift the first value from the "values" property in the "resolution" object from the LUIS response.