Bumps version to 0.4.0
The next release contains a breaking change for the `Entity` DTO, where we removed the `EntityResolution` property. As such, in addition to bumping the version on `NLU.DevOps.CommandLine`, we are also bumping the version on `NLU.DevOps.Models` and `NLU.DevOps.Core`. Bumps each version to 0.4.0 in preparation for release.