Runs NUnit tests for compare command in parallel.
The bottleneck for the `compare` command is the exceptions thrown by NUnit for `Assert.Pass` and `Assert.Fail`. Running the tests in parallel will leverage more threads to run the tests. Another thing we can look into if we want to improve this further is to remove the calls to `Assert.Pass` and `Assert.Fail` and instead use expected outputs for unit tests, to remove the need for exception handling. For now, this parallel approach dramatically increases throughput for compare tests.

Fixes #138