Resolves LUIS v2 entities with roles
LUIS v3 sets the entity type for entities with roles to the role name. This change adds a similar resolution mechanism to LUIS v2, resolving entities with a "role" property to use the role for the entity type. This change also supports training LUIS v2 and v3 models with entities labeled with roles by adding the "roles" property to the `LuisSettings` supplied with the `--model-settings` CLI option.

Fixes #166