Adds JsonExtensionData to labeled utterances
Adds `AdditionalData` property annotated with the `JsonExtensionDataAttribute` to capture additional JSON beyond the current `LabeledUtterance` data model when reading in utterances for the `train` command. This will allow us to create a demo with the `NLU.DevOps.MockProvider` that shows how multi-intent scenarios can be handled. Note that we are not currently exposing the `AdditionalData` property as a public API, at least not until another scenario warrants this.

Fixes #183