Adds support for dynamic entity value matching
In order to support matching of dynamic entity values, this commit introduces to changes.

First, it adds a `context` property to each predicted utterance from the NLU service, including a `timestamp` value for when the prediction request was made.

Second, it adds a mechanism to include inline scripts, which are evaluated at runtime with Roslyn.

Fixes #204