Adds configuration to clean version only
We've run into a few cases where the `clean` command inadvertently deletes that was not created within the scope of a single pipeline run, and this was not the desired behavior.

This commit adds functionality so that the LUIS app is only deleted when the `luisAppCreated` configuration value is set to `true`, which will be the case when a new LUIS app is created by the NLU.DevOps tool and the `--save-appsettings` flag is used for training.

Fixes #242