Simplifying LUIS configuration
Previously, we supported a custom configuration that allowed you to specify mappings for prebuilt entity types and also hooks to allow roles to be used in place of entity names.

This change removes the prebuilt entity mappings and changes the behavior so that, during training, if you specify a role name as an entity name and have that role defined in the `--model-settings` template provided, the entity will be sent to LUIS with the correct role and entity type information.

Fixes #262