Adds option to set absolute thresholds
For the performance regression functionality in the `compare` command, this change adds the option to set absolute thresholds for F<sub>1</sub> scores. For example, a user may want to ensure that the F<sub>1</sub> score for a given intent never falls below 0.5. To define this, with this change, they could add a threshold as follows:
```yaml
thresholds:
- type: intent
  group: PlayMusic
  comparison: absolute
  threshold: 0.5
```