Adds option to customize the regression metric (#314)
This change allows customization of the metric used to check for performance regressions. For example, if a user wants to set a threshold for regression targeted only at precision, recall or a custom F-score, they can add the `metric` property to their threshold configuration in the compare test settings. The valid values are `precision`, `recall` or `f(\d+)(.\d+)?` (e.g., `f0.5`).

Fixes #313