Adds support for running from .NET tool manifest
Previously, the service resolver only worked when an include path was provided or the .NET CLI tool for dotnet-nlu was either installed to a tool path or globally. This change adds support for running the dotnet-nlu CLI tool from a local tool configured in a tool manifest. See [documentation](https://docs.microsoft.com/en-us/dotnet/core/tools/local-tools-how-to-use) for details.

Fixes #324